refactor(features): вынес delete accoount в отдельный компонент
