+++ src/entities/delete-account/index.ts
@@ -1,0 +1,1 @@
+export { DeleteAccount } from './ui/delete-account';
