+++ src/entities/delete-account/ui/types.ts
@@ -1,0 +1,3 @@
+export type DeleteAccountProps = {
+  handleDeleteAccount: (currentPassword: string) => void;
+};
