+++ src/entities/form-profile-settings/ui/form-profile-settings.tsx
@@ -2,9 +2,9 @@
 
 import React, { FC, useState } from 'react';
 // import { Controller, useForm } from 'react-hook-form';
-import { Input, MainButton } from '@/shared/ui';
+import { MainButton } from '@/shared/ui';
 import { Form } from '@/shared/ui';
-import { PopUp } from '@/shared/ui/pop-up/pop-up';
+// import { PopUp } from '@/shared/ui/pop-up/pop-up';
 import IconUp from '@/shared/assets/icons/chevron-up.svg';
 import IconDown from '@/shared/assets/icons/chevron-down.svg';
 import styles from './form-profile-settings.module.scss';
@@ -15,34 +15,33 @@ import { FormProfileSettingsProps } from './types';
 
 export const FormProfileSettings: FC<FormProfileSettingsProps> = ({
 	handleSubmitForm,
-	handleDeleteAccount,
+	// handleDeleteAccount,
 	// settingsProfile,
 }) => {
 	// const methods = useForm();
 	// const { control, handleSubmit } = methods;
 
-	const [isPopupOpen, setIsPopupOpen] = useState(false);
 	const [isSendNotification, setIsSendNotification] = useState(false);
 	const [isSubscriptionProjects, setIsSubscriptionProjects] = useState(false);
 	const [showVisibleProfileMenu, setShowVisibleProfileMenu] = useState(false);
 	const [showVisibleContactsMenu, setShowVisibleContactsMenu] = useState(false);
-	const [currentPasswordValue, setCurrentPasswordValue] = useState('');
+	// const [currentPasswordValue, setCurrentPasswordValue] = useState('');
 
-	const handleOpenPopup = () => setIsPopupOpen(true);
+	// const handleOpenPopup = () => setIsPopupOpen(true);
 
-	const handleClosePopup = () => setIsPopupOpen(false);
+	// const handleClosePopup = () => setIsPopupOpen(false);
 
-	const handleChangeInput = (
-		evt: React.ChangeEvent<HTMLInputElement>
-	): void => {
-		evt.preventDefault();
-		setCurrentPasswordValue(evt.target.value);
-	};
-	const handleDeleteAccountAction = () => {
-		handleDeleteAccount(currentPasswordValue);
-		setCurrentPasswordValue('');
-		// setIsPopupOpen(false);
-	};
+	// const handleChangeInput = (
+	// 	evt: React.ChangeEvent<HTMLInputElement>
+	// ): void => {
+	// 	evt.preventDefault();
+	// 	setCurrentPasswordValue(evt.target.value);
+	// };
+	// const handleDeleteAccountAction = () => {
+	// 	handleDeleteAccount(currentPasswordValue);
+	// 	setCurrentPasswordValue('');
+	// 	// setIsPopupOpen(false);
+	// };
 
 	return (
 		<section className={styles.profileSettings}>
@@ -156,12 +155,7 @@ export const FormProfileSettings: FC<FormProfileSettingsProps> = ({
 						</div>
 					</div>
 				</div>
-				<button
-					type="button"
-					className={styles.formSettings__deleteButton}
-					onClick={handleOpenPopup}>
-					Удалить аккаунт
-				</button>
+
 				<MainButton
 					// type="button"
 					variant={'primary'}
@@ -170,38 +164,6 @@ export const FormProfileSettings: FC<FormProfileSettingsProps> = ({
 				>
 					Сохранить настройки
 				</MainButton>
-				<PopUp
-					visible={isPopupOpen}
-					title="Удаление аккаунта"
-					onClose={handleClosePopup}>
-					<div className={styles.deleteAccount}>
-						<Input
-							className={styles.deleteAccount__input}
-							onChange={handleChangeInput}
-							name="repeatNewPassword"
-							type="password"
-							labelName="Для удаления введите пароль"
-							value={currentPasswordValue}
-						/>
-						<div className={styles.deleteAccount__buttons}>
-							<MainButton
-								type="button"
-								variant={'primary'}
-								disabled={currentPasswordValue.length < 8}
-								width={'regular'}
-								onClick={handleDeleteAccountAction}>
-								Удалить аккаунт
-							</MainButton>
-							<MainButton
-								type="button"
-								variant={'secondary'}
-								width={'regular'}
-								onClick={handleClosePopup}>
-								Вернуться в настройки
-							</MainButton>
-						</div>
-					</div>
-				</PopUp>
 			</Form>
 		</section>
 	);
