+++ src/entities/form-profile-settings/ui/types.ts
@@ -1,6 +1,5 @@
 
 export type FormProfileSettingsProps = {
-  handleDeleteAccount: (currentPassword: string) => void;
   handleSubmitForm: (data:unknown) => void;
   // settingsProfile: unknown;
 }
