+++ src/features/index.ts
@@ -3,5 +3,6 @@ export { FormSignupConfirmFeature } from '@/features/form-signup-confirm/form-si
 export { FormLoginFeature } from '@/features/form-login/form-login';
 export { FormPasswordRecoveryFeature } from '@/features/form-password-recovery/form-password-recovery';
 export { LikeButtonFeature } from '@/features/like-button/like-button';
-export { FormChangePasswordFeature } from '@/features/form-change-password/form-change-password'
-export { FormProfileSettingsFeature } from '@/features/form-profile-settings/form-profile-settings'
+export { FormChangePasswordFeature } from '@/features/form-change-password/form-change-password';
+export { FormProfileSettingsFeature } from '@/features/form-profile-settings/form-profile-settings';
+export { DeleteAccountFeature } from '@/features/delete-account/delete-account';
